backend/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
pytest>=8.0,<9.0
httpx>=0.27,<1.0
numpy>=2.0,<3.0
Pillow>=10.0,<12.0
opencv-python-headless>=4.10,<5.0
python-multipart>=0.0.9,<1.0
